fix qos_time()

+++ time.cpp
@@ -3,7 +3,7 @@
 #include "hardware/timer.h"
 
 qos_time_t qos_time() {
-  return time_us_64() + INT_MIN;
+  return time_us_64() + INT64_MIN;
 }
 
 void qos_normalize_time(qos_time_t* time) {
